requirements.txt
gitpython>=3.1.0
transformers>=4.30.0
torch>=2.0.0
pyyaml>=6.0.1